Lots of changes and updates
Finished with BakeryGoodController for now. Bakery goods can be added and deleted to the list view and gen linked list.
Started on the basics @FXML parts in IngredientController.
Cleaned up BakeryGoods.fxml code and renamed it to BakeryGood.fxml BECAUSE  R E A S O N S.
Added some methods to the GenList to facilitate further controller functions etc.