Updated 17 files
Quite a few minor alterations here and there throughout the project, main piece of work went into the recipeController and Recipe.fxml, I have it so that the you must choose a bakery good, then select an ingredient from the ingredient's list then clicking a button to add the ingredient prompts you to enter a desired quantity, then once enter is clicked it adds it to the chosenIngredientsListView along with the quantity, as a temporary fix due to the complicated way handling listviews and object types works i made chosenIngredients listview of type stringand the to be added listview as type ingredient, Jordan mentioned earlier about making a hidden variable in ingredient class for quantity, i have implemented something along these lines that is hidden untill its needed when its set by the user when making the recipe, in order to do this i created a second to string to store the string object in the list view. Complicated stuff ahead added the next major TODO for clarification near the addRecipe method which is going to be complicated to implement.